requirements.txt
# Podstawowe biblioteki numeryczne i ML
numpy~=1.24.0
pandas==1.5.3
scipy==1.11.4
scikit-learn==1.5.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.5.0
threadpoolctl==3.6.0
kiwisolver==1.4.8
contourpy==1.3.2
cycler==0.12.1
pyparsing==3.0.9
fonttools==4.58.0
pillow==11.2.1

# Biblioteki NLP
spacy~=3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
thinc~=8.3.6
cymem==2.0.11
preshed==3.0.9
murmurhash==1.0.12
blis~=1.3.0
catalogue==2.0.10
srsly==2.5.1
wasabi==1.1.3
langcodes==3.5.0
language_data==1.3.0
pathy==0.11.0
typer~=0.15.4
smart-open~=7.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.6.0/pl_core_news_sm-3.6.0-py3-none-any.whl#sha256=6c7403e7e124d55b672d1a418f7e0fa9c388cb804e000c9ad315cf317837eb27

# Web API i serwery
Flask~=3.1.1
Werkzeug~=3.1.3
click~=8.1.8
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
itsdangerous==2.2.0

# Baza danych
PyMySQL==1.1.1

# Qdrant Vector Database
qdrant-client==1.12.1
sentence-transformers==3.4.0
grpcio==1.71.1
grpcio-tools==1.71.1
portalocker==3.0.1
huggingface-hub==0.28.1
tokenizers==0.21.0
transformers==4.48.1
safetensors==0.5.2
filelock==3.17.0
fsspec==2025.1.0
regex==2025.1.24

# HTTP i Networking
aiohttp==3.11.18
aiosignal==1.3.2
aiohappyeyeballs==2.6.1
requests==2.32.3
httpx==0.27.0
httpcore==1.0.9
openai==1.23.6
h11==0.16.0
sniffio==1.3.1
anyio==4.9.0
yarl==1.20.0
multidict==6.4.3
frozenlist==1.6.0
charset-normalizer==3.4.2
certifi==2025.4.26
urllib3==2.4.0
idna==3.10

# Utilities i pomocnicze
python-dateutil==2.8.2
pytz==2025.2
tzdata==2025.2
six==1.17.0
tqdm==4.67.1
reportlab==4.0.4
schedule~=1.2.2
ffmpeg-python==0.2.0
pathlib_abc==0.1.1
discord.py==2.3.2
distro==1.9.0
colorama==0.4.6
future==1.0.0
typing_extensions==4.13.2
typing-inspection==0.4.0
annotated-types==0.7.0
python-dotenv==1.0.0
confection==0.1.5
propcache==0.3.1
marisa-trie==1.2.1
attrs==25.3.0
packaging==25.0

# Wersja pydantic dopasowana do starszej wersji NumPy
pydantic~=2.11.4
flask-cors~=6.0.0
pip~=24.0
rich~=14.0.0
typing_extensions~=4.13.2
Pygments~=2.19.1
setuptools~=70.0.0
shellingham~=1.5.4
pydantic_core~=2.33.2
language_data~=1.3.0